requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.0.1
boto2==0.1.0
boto3==1.37.38
botocore==1.37.38
cffi==1.17.1
click==8.1.8
cryptography==46.0.4
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
idna==3.11
jmespath==1.0.1
passlib==1.7.4
pkg_resources==0.0.0
pyasn1==0.4.8
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
rsa==4.9.1
s3transfer==0.11.5
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.33.0
